requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
openai
pydantic
pydantic-settings
jinja2
